+++ graphql/openreader/src/model.schema.ts
@@ -28,7 +28,7 @@ const baseSchema = buildASTSchema(parse(`
     directive @query on INTERFACE
     directive @derivedFrom(field: String!) on FIELD_DEFINITION
     directive @unique on FIELD_DEFINITION
-    directive @index(fields: [String!] unique: Boolean) on OBJECT | FIELD_DEFINITION
+    directive @index(fields: [String!] unique: Boolean) repeatable on OBJECT | FIELD_DEFINITION
     directive @fulltext(query: String!) on FIELD_DEFINITION
     directive @cardinality(value: Int!) on OBJECT | FIELD_DEFINITION
     directive @byteWeight(value: Float!) on FIELD_DEFINITION