+++ substrate/substrate-metadata/src/old/definitions/orml.ts
@@ -78,5 +78,5 @@ export const ormlAlias: OldTypesAlias = {
     tokens: {
         AccountData: 'OrmlAccountData',
         BalanceLock: 'OrmlBalanceLock'
-      }
+    }
 }