+++ substrate/ink-abi/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log - @subsquid/ink-abi
 
-This log was last generated on Sun, 06 Nov 2022 18:24:20 GMT and should not be manually modified.
+This log was last generated on Wed, 14 Dec 2022 19:37:27 GMT and should not be manually modified.
+
+## 0.0.3
+Wed, 14 Dec 2022 19:37:27 GMT
+
+### Patches
+
+- fix invalid name generation for messages with '::'
 
 ## 0.0.2
 Sun, 06 Nov 2022 18:24:20 GMT